add schemas and fix route transaction
