+++ models/transactionModel.js
@@ -1,5 +1,6 @@
 const mongoose = require("mongoose");
-const Joi = require("joi");
+
+const { transactionSchema } = require("../schemas/transaction");
 
 const schema = mongoose.Schema(
   {
@@ -38,15 +39,6 @@ const schema = mongoose.Schema(
 
 const Transaction = mongoose.model("transaction", schema);
 
-const transactionSchema = Joi.object({
-  date: Joi.date().required(),
-  description: Joi.string().required(),
-  category: Joi.string().required(),
-  sum: Joi.number().required(),
-  operation: Joi.string().required(),
-});
+const schemas = { transactionSchema };
 
-module.exports = {
-  Transaction,
-  transactionSchema,
-};
+module.exports = { Transaction, schemas };
