+++ schemas/user.js
@@ -1,0 +1,9 @@
+const Joi = require("joi");
+
+const balanceSchema = Joi.object({
+    totalBalance: Joi.number().required(),
+});
+
+module.exports = {
+    balanceSchema,
+};
